auto_labeling/pyproject.toml
[tool.poetry]
name = "auto_labeling"
version = "0.1.0"
description = ""
authors = ["Danil Zlobin <xzlobin@stuba.sk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"

# workaround to solve cv2.imshow freezing after import torchvision (from phycv)
# https://github.com/pytorch/vision/issues/5940
# https://github.com/opencv/opencv/issues/21952
av = "<9"

numpy = "^1.26.4"
phycv = "^1.2.6"
pandas = "^2.2.1"
pygad = "^3.3.1"
typing-extensions = "^4.10.0"
nptyping = {extras = ["complete"], version = "^2.5.0"}
opencv-python = "^4.9.0.80"
torchvision = "^0.17.2"
imutils = "^0.5.4"
scikit-learn = "^1.4.2"
seaborn = "^0.13.2"
pipe = "^2.2"
kneed = "^0.8.5"
subpixel-edges = "^0.1.1"
tqdm = "^4.66.4"
tensorflow = "^2.16.1"
torchmetrics = "^1.4.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
